requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.6
django-environ==0.4.5
idna==3.2
pytz==2021.1
requests==2.26.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.6
